+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapClient.lua
@@ -1,5 +1,5 @@
 --  #################################################################################################################################
---  #########                                       STATS CONFIGURATOR - BOOTSTRAPCLIENT                                    #########
+--  #########                                       STATS CONFIGURATOR - BOOTSTRAP_CLIENT                                   #########
 --  =================================================================================================================================
 Ext.Require("S7_ConfigAuxiliary.lua")
 Ext.Require("S7_ConfigCollections.lua")
@@ -14,7 +14,7 @@ logSource = "Lua:BootstrapClient"
 
 local function StatsLoader() --  Loads stats-configuration json after StatsLoaded Event.
     if ConfigSettings.StatsLoader.Enable == true then --  StatsLoader enabled in settings.
-        S7_ConfigLog("Loading " .. ConfigSettings.StatsLoader.FileName)
+        S7_ConfigLog("StatsLoader active. Loading " .. ConfigSettings.StatsLoader.FileName)
         local file = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) or "" --  Load file if it exists. Load empty string otherwise.
         if ValidJSONFile(file) then --  if configData file exists and is not empty.
             local configData = Ext.JsonParse(file) --  Parse into table.
@@ -29,7 +29,7 @@ local function StatsLoader() --  Loads stats-configuration json after StatsLoade
             StatsConfigurator() --  Configure Stats
             toConfigure = {} -- flush config queue
             toSync = {} --  flush Sync queue
-            S7_ConfigLog("StatsLoading Completed.")
+            S7_ConfigLog("StatsLoading completed.")
         else
             S7_ConfigLog("Failed to load " .. ConfigSettings.StatsLoader.FileName, "[Error]")
         end
@@ -48,16 +48,21 @@ Ext.RegisterListener("StatsLoaded", StatsLoader)
 
 local function CatchBroadcast(channel, payload) --  Listens for broadcasts from the Server.
     if channel == "S7_ConfigData" then --  if broadcast channel is S7_ConfigData.
-        S7_ConfigLog("Client recieved Active Configuration. Saving File: " .. ConfigSettings.StatsLoader.FileName)
+        S7_ConfigLog("Client recieved active configuration. Saving file: " .. ConfigSettings.StatsLoader.FileName)
         Ext.SaveFile(ConfigSettings.StatsLoader.FileName, payload) --  Save stringified json.
     end
 
     if channel == "S7_ValidateClientConfig" then --  if broadcast channel is S7_ValidateClientConfig
-        local verify = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) or "" --    Load local ActiveConfiguration if available
-        if payload == verify then
-            Ext.PostMessageToServer("S7_ValidateClientResponse", "All Good.")
-        else
-            Ext.PostMessageToServer("S7_ValidateClientResponse", "Config Mismatch Detected.")
+        local verify = Ext.LoadFile(ConfigSettings.StatsLoader.FileName) or "" --    Load local ActiveConfiguration if available.
+        for clientID, compare in pairs(Ext.JsonParse(payload)) do
+            local message = clientID .. " : "
+            if ValidJSONFile(verify) and compare == verify then
+                message = message .. "Active configuration profile verified."
+                Ext.PostMessageToServer("S7_ValidateClientResponse", message)
+            else
+                message = message .. "Active configuration mismatch detected."
+                Ext.PostMessageToServer("S7_ValidateClientResponse", message)
+            end
         end
     end
 end