+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 * _Moved_ user information stuff in `S7_ConfigAuxiliary.lua` from `S7_ConsoleCommander.lua`.
 * `S7_InspectStats` **Reborn**.
 * `S7_ExportStatIDtoTSV` simplified.
-* _Dropped_ prefix from `table ConfigSettings`, `table DefaultSettings`, `func StatsLoader`, `func CatchBroadcast`, `func ValidJSONFile`.
+* _Dropped_ prefix from `table ConfigSettings`, `table DefaultSettings`, `func StatsLoader`, `func CatchBroadcast`, `func ValidJSONFile` and more. You get the idea.
+* _Revamped_ networking.
 
 ## [0.6.1.0] --- 28th June 2020 --- **_Message-Boxes!_**